+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased 0.0.3
 
 - Decrease refresh delay
+- Support multiple workspace folders
 
 ## [0.0.2](https://github.com/shoorick/codeception-test-adapter/releases/tag/v0.0.2) - 2025-12-17